Moved getNewlySelectedKeys spec to helper, and updated for select-all function
